ANM-1: New file for role in backend
