+++ backend/src/index.ts
@@ -3,6 +3,7 @@ import express from "express";
 import bodyparser from "body-parser";
 import authRoute from "./routes/authRoute";
 import adminRoute from "./routes/adminRoute";
+import roleRoute from "./routes/roleRoute";
 import cors from "cors";
 import { DB, DBHOST, DBPORT, PORT } from "./config/env-vars";
 
@@ -20,3 +21,4 @@ app.listen(PORT, () => console.log("Server run on 8080"));
 
 app.use("/api", authRoute);
 app.use("/api", adminRoute);
+app.use("/api", roleRoute);
