+++ frontend/src/app/role/role.service.ts
@@ -2,6 +2,8 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Role, RoleLevel, RoleResponse } from '../interfaces/role.interface';
+import { environment } from 'src/environments/environment';
+const API_BASE = environment.apiBase;
 
 @Injectable({
   providedIn: 'root',
@@ -10,22 +12,22 @@ export class RoleService {
   constructor(private http: HttpClient) {}
 
   createRole(body: Role): Observable<RoleResponse> {
-    return this.http.post<RoleResponse>('http://localhost:8080/api/createRole', body);
+    return this.http.post<RoleResponse>(`${API_BASE}/createRole`, body);
   }
 
   updateRole(body: Role, id: string): Observable<RoleResponse> {
-    return this.http.put<RoleResponse>(`http://localhost:8080/api/updateRole/${id}`, body);
+    return this.http.put<RoleResponse>(`${API_BASE}/updateRole/${id}`, body);
   }
 
   deleteRole(body: Role, id: string): Observable<RoleResponse> {
-    return this.http.put<RoleResponse>(`http://localhost:8080/api/deleteRole/${id}`, body);
+    return this.http.put<RoleResponse>(`${API_BASE}/deleteRole/${id}`, body);
   }
 
   getAllRoles(body: {}): Observable<Role[]> {
-    return this.http.post<Role[]>(`http://localhost:8080/api/getAllRoles`, body);
+    return this.http.post<Role[]>(`${API_BASE}/getAllRoles`, body);
   }
 
-  getAllRoleLevels(): Observable<RoleLevel[]>{
-    return this.http.get<RoleLevel[]>('http://localhost:8080/api/getAllRoleLevels');
+  getAllRoleLevels(): Observable<RoleLevel[]> {
+    return this.http.get<RoleLevel[]>(`${API_BASE}/getAllRoleLevels`);
   }
 }
